requirements.txt
gensim==4.2.0
hydra-core==1.2.0
matplotlib==3.5.2
nltk==3.7
numpy==1.22.4
omegaconf==2.2.2
packaging==21.3
pandas==1.4.2
pytest==7.1.2
python-dotenv==0.20.0
rich==12.4.4
sh==1.14.2
torch==2.4.1
torch_geometric==2.0.4
torchmetrics==0.9.1
transformers==4.19.4
contractions==0.1.72
pytorch-lightning
torch_sparse
hydra-colorlog
hydra-optuna-sweeper
wandb
setuptools